--- MANIFEST.in
@@ -1,2 +1,0 @@
-recursive-include jupyter-config *.json
-recursive-include data-grid *.js
