+++ pyproject.toml
@@ -10,50 +10,28 @@ readme = "README.md"
 requires-python = ">=3.13"
 license = {file = "LICENSE"}
 dependencies = [
-    "ipywidgets>=8.1.7",
-    "traitlets>=5.14.3",
+    "anywidget>=0.9.18",
 ]
 
 [dependency-groups]
 dev = [
-    "ipywidgets>=8.1.7",
-    "jupyter>=1.1.1",
+    "jupyterlab>=4.4.9",
     "jupyterlite-core>=0.6.4",
     "jupyterlite-pyodide-kernel>=0.6.1",
     "pandas>=2.3.3",
+    "watchfiles>=1.1.0",
 ]
 
-[tool.hatch.build.targets.sdist]
-artifacts = ["data_grid/labextension"]
-exclude = [".github", "binder"]
-
-[tool.hatch.build.targets.wheel.shared-data]
-"data_grid/labextension" = "share/jupyter/labextensions/data-grid"
-"install.json" = "share/jupyter/labextensions/data-grid/install.json"
+[tool.hatch.build]
+only-packages = true
+artifacts = ["src/data_grid/static/*"]
 
 [tool.hatch.build.hooks.jupyter-builder]
-dependencies = ["hatch-jupyter-builder>=0.5"]
 build-function = "hatch_jupyter_builder.npm_builder"
-ensured-targets = [
-    "data_grid/labextension/static/style.js",
-    "data_grid/labextension/package.json",
-]
-skip-if-exists = ["data_grid/labextension/static/style.js"]
+ensured-targets = ["src/data_grid/static/widget.js"]
+skip-if-exists = ["src/data_grid/static/widget.js"]
+dependencies = ["hatch-jupyter-builder>=0.5.0"]
 
 [tool.hatch.build.hooks.jupyter-builder.build-kwargs]
-build_cmd = "build:prod"
-npm = ["jlpm"]
-
-[tool.jupyter-releaser.options]
-version_cmd = "hatch version"
-
-[tool.jupyter-releaser.hooks]
-before-build-npm = [
-    "python -m pip install 'jupyterlab>=4.0.0,<5'",
-    "jlpm",
-    "jlpm build:prod"
-]
-before-build-python = ["jlpm clean:all"]
-
-[tool.check-wheel-contents]
-ignore = ["W002"]
+npm = "npm"
+build_cmd = "build"
