--- style/index.css
@@ -1,1 +1,0 @@
-@import url('base.css');
